+++ apps/golden-sample-app-e2e/data/mocks-data/index.ts
@@ -1,0 +1,1 @@
+export * from './transactions.mocks.api.data';
