+++ apps/golden-sample-app-e2e/data/sandbox-api-data/index.ts
@@ -1,0 +1,1 @@
+export * from './transactions.sandbox.api.data';
