+++ apps/golden-sample-app-e2e/specs/mb-login.spec.ts
